test(reborn): extract host runtime services test harness (#3231)
host_runtime_services_contract.rs mixed 109 vertical-slice tests with
~2.7k lines of shared setup (fixtures, fake/recording backends, manifest
constants, assertion and scope/context/id builder helpers), so each test
read as a setup script rather than a domain scenario.

Move all non-test helper items behind a support module,
tests/support/host_runtime_harness.rs, exposed as pub(crate) and imported
via `use support::host_runtime_harness::*;`. Behavior-preserving: every
test body, assertion, and test name is unchanged; only repeated setup
moved. The contract file drops from 8828 to 6217 lines.

Note: the WIT fixture helper moved one directory deeper, so its
include_str! path changed from ../../../wit/tool.wit to
../../../../wit/tool.wit (still resolves to repo wit/tool.wit).

Implements item 1 of #3231.

Verified: cargo test -p ironclaw_host_runtime (default and --features
libsql), the host_runtime_services_contract binary (104 default / 109
libsql), cargo clippy --all-targets -- -D warnings (both feature sets),
cargo fmt --check, and git diff --check all pass.
