+++ crates/ironclaw_host_runtime/tests/support/mod.rs
@@ -1,5 +1,7 @@
 #![allow(dead_code)]
 
+pub(crate) mod host_runtime_harness;
+
 pub(crate) fn legacy_capability_fixture_to_v2(manifest: &str) -> String {
     legacy_capability_fixture_to_v2_with_refs(manifest, |_| {
         (
